Don't leak message objects through response_map. (#14)
Add get/setMID to Response objects.
Add SmbTransportInternal.getInflighRequests() for testing.

+++ src/main/java/jcifs/internal/smb1/trans/nt/SmbComNtTransaction.java
@@ -61,7 +61,7 @@ public abstract class SmbComNtTransaction extends SmbComTransaction {
      */
     @Override
     public CommonServerMessageBlockRequest createCancel () {
-        return new SmbComNtCancel(getConfig(), getMid());
+        return new SmbComNtCancel(getConfig(), (int) getMid());
     }
